created custom hook for playerstats
