+++ pages/_app.js
@@ -3,7 +3,7 @@ import GlobalStyle from "../styles";
 
 const fetcher = (...args) => fetch(...args).then((res) => res.json());
 
-const CURRENT_SEASON = 2022;
+const CURRENT_SEASON = "2022";
 
 export default function App({ Component, pageProps }) {
   return (
